raphael: Update libfod extension function prototypes

+++ fod/FodExtension.cpp
@@ -25,6 +25,6 @@ uint32_t getFodZOrder(uint32_t z, bool touched) {
     return z;
 }
 
-uint32_t getFodUsageBits(uint32_t usageBits, bool) {
+uint64_t getFodUsageBits(uint64_t usageBits, bool) {
     return usageBits;
 }
